% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhd.line.mo}
\alias{nhd.line.mo}
\title{NHD Line data cropped to the state of Missouri}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1150 rows and 14 columns.
}
\usage{
nhd.line.mo
}
\description{
NHD data obtained throught the FedData package in R
}
\details{
see the NHD data dictionary for for further detail:
https://www.usgs.gov/ngp-standards-and-specifications/national-hydrography-dataset-nhd-data-dictionary-feature-domains
}
\keyword{datasets}
